+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Handle `Agent` tool type as an alias for `Task`
 - Handle display of TodoWrite tool
 - Syntax-highlighted code display for Read tool results
+- Show URL for WebFetch and search query for WebSearch in tool use summaries
 
 ## [1.0.3] - 2026-03-04